circle refactoring complete
