+++ geometry/__init__.py
@@ -1,2 +1,3 @@
 from .point import Point
 from .line import Line
+from .circle import Circle
